Optimize OCR Receipt Build
