+++ frontend/app/components/elements/CodeSnippet/CodeSnippet.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import SyntaxHighlighter from 'react-syntax-highlighter'
+import { LightAsync as SyntaxHighlighter } from 'react-syntax-highlighter'
 import { googlecode } from 'react-syntax-highlighter/dist/cjs/styles/hljs'
 import Image from 'next/image'
