+++ frontend/pages/innovations/ocr-receipt.tsx
@@ -1,14 +1,7 @@
 import { getServiceBySlug } from '@/api/analyticsAPI'
-import { CodeSnippet } from '@/elements/CodeSnippet/CodeSnippet'
-import { ReceiptDisplay } from '@/modules/ReceiptDisplay/ReceiptDisplay'
-import { AnalyticsPage } from '@/templates/AnalyticsPage/AnalyticsPage'
 import { ServiceBySlugResponseData } from '@/types/responses'
-import { isOCRReceipt } from '@/utils/utils'
-import { WarningDiv } from '@/elements/WarningDiv/WarningDiv'
 import Head from 'next/head'
-import { useState } from 'react'
-import { postInnovation } from './../../app/api/innovationsAPI'
-import { MAX_IMAGE_SIZE } from 'app/constants/constant'
+import { OCRReceiptPage } from '@/templates/OCRReceiptPage/OCRReceiptPage'
 
 const OCRReceipt = ({
   name,
@@ -16,51 +9,17 @@ const OCRReceipt = ({
   id,
   slug
 }: ServiceBySlugResponseData) => {
-  const [res, setRes] = useState()
-  const renderResult = () => {
-    // Error message passed
-    if (typeof res !== 'undefined') {
-      if (typeof res === 'string') {
-        return <WarningDiv message={res} />
-      }
-      if (isOCRReceipt(res)) {
-        return <ReceiptDisplay result={res} />
-      } else {
-        return <CodeSnippet lang="json" code={JSON.stringify(res, null, 3)} />
-      }
-    } else {
-      return (
-        <div>
-          {'There is a problem on our analytics. Please use another photo.'}
-        </div>
-      )
-    }
-  }
-
   return (
     <>
       <Head>
         <title>{`Innovation | ${name} - Demo`}</title>
       </Head>
-      <AnalyticsPage
-        analyticsName={name}
-        longDescription={long_description}
-        examples={[
-          `/assets/images/innovations/${slug}/example1.jpg`,
-          `/assets/images/innovations/${slug}/example2.jpg`,
-          `/assets/images/innovations/${slug}/example3.jpg`
-        ]}
-        serviceID={id}
+      <OCRReceiptPage
+        id={id}
+        long_description={long_description}
+        name={name}
         slug={slug}
-        handlePost={(session_id, photo) =>
-          postInnovation(id, session_id, photo)
-        }
-        handleResult={(res) => setRes(res)}
-        maxImageSize={MAX_IMAGE_SIZE}
-        acceptedFileFormat={'image/jpeg, image/png'}
-        customBannerUrl={'/assets/images/innovations/ocr-receipt/banner.png'}>
-        {renderResult()}
-      </AnalyticsPage>
+      />
     </>
   )
 }
